pyproject.toml
[project]
name = "ebay-kleinanzeigen-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "playwright>=1.49.0",
    "pytest>=8.4.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "uvicorn>=0.34.0",
    "uvloop>=0.21.0",
    "imagehash>=4.3.1",
    "pillow>=11.0.0",
    "numpy>=2.1.3",
    "prometheus-client>=0.21.0",
    "tzdata>=2024.1",
    "pgeocode>=0.5.0",
    "pandas>=2.2.3",
    "requests>=2.32.3",
]
